wip: cleanup connectors cache usage
