+++ src/platform/plugins/shared/workflows_management/common/connectors_contracts/enhanced_es_connectors.ts
