+++ src/platform/plugins/shared/workflows_management/common/connectors_contracts/stack_connectors_schema/resilient.ts
