+++ src/platform/plugins/shared/workflows_management/public/widgets/workflow_yaml_editor/lib/autocomplete/get_completion_item_provider.ts
@@ -40,7 +40,6 @@ export function getCompletionItemProvider(
       }
 
       const suggestions = getSuggestions({ ...autocompleteContext, model, position });
-
       return {
         suggestions,
         incomplete: false,
